+++ src/pages/Callback/style.scss
@@ -1,0 +1,6 @@
+.callback {
+    width: 100vw;
+    height: 100vh;
+    background-color: #ffffff;
+    border: 1px solid black;
+}
